initialize attribute currentUser and add a null check in tear down to avoid dependence of assert
git-svn-id: http://vcs.taotesting.com/svn/taoqual/branches/WorlflowEngineExt@1420 793250d4-a296-4af0-ab9f-c080ea916244

+++ test/NotificationServiceTestCase.php
@@ -69,7 +69,9 @@ class NotificationServiceTestCase extends UnitTestCase {
 	}
 	
 	public function tearDown() {
-		$this->currentUser->delete();
+		if (!is_null($this->currentUser)){
+			$this->currentUser->delete();
+		}
     }
 	
 	/**
